requirements.txt
pandas
pymongo
slack_sdk
Flask
requests
python-dotenv
numpy